gopy, copyToFolder, mkdir

+++ tests/core/Sagres/Framework/FileSystem/ActionTest.php
@@ -42,10 +42,12 @@ class ActionTest extends \PHPUnit_Framework_TestCase
 
     private function resetFolder($folder) {
 
-        if( $dh = opendir($folder)) {
-            while (($filename = readdir($dh)) !== false) {
-                if(! in_array($filename, array('.', '..'))) {
-                    unlink($folder . DIRECTORY_SEPARATOR . $filename);
+        if(is_dir($folder)) {
+            if( $dh = opendir($folder)) {
+                while (($filename = readdir($dh)) !== false) {
+                    if(! in_array($filename, array('.', '..'))) {
+                        unlink($folder . DIRECTORY_SEPARATOR . $filename);
+                    }
                 }
             }
         }
@@ -58,13 +60,16 @@ class ActionTest extends \PHPUnit_Framework_TestCase
 
     /**
      * @covers Sagres\Framework\FileSystem\Action::copyToFolder
+     * @covers Sagres\Framework\FileSystem\Action::copy
+     * @covers Sagres\Framework\FileSystem\Action::getPermissions
+     * @covers Sagres\Framework\FileSystem\Action::getDestination
      */
     public function testCopyToFolder()
     {
         $fileSet = new Set();
-        $fileSet->addSet(__DIR__ . '/../../../../fixtures/folder1');
+        $fileSet->addSet(__DIR__ . '/../../../../fixtures/folder1/');
 
-        $folder = __DIR__ . '/../../../../fixtures/copy';
+        $folder = __DIR__ . '/../../../../fixtures/copy/';
 
         // make sure the folder is empty
         $this->resetFolder($folder);
@@ -80,14 +85,15 @@ class ActionTest extends \PHPUnit_Framework_TestCase
 
     /**
      * @covers Sagres\Framework\FileSystem\Action::copyToFolder
-     * @expectedException Sagres\Framework\FileSystem\Exception\notFound
+     * @covers Sagres\Framework\FileSystem\Action::copy
+     * @expectedException \LogicException
      * @depends testCopyToFolder
      */
-    public function testCopyToFolder_folderNotFound()
+    public function testCopyToFolder_noCommonBaseSourceFolder()
     {
         $fileSet = new Set();
         $fileSet->addSet(__DIR__ . '/../../../../fixtures/folder1');
-
+        $fileSet->addPath('/var/');
         $folder = __DIR__ . '/../../../../fixtures/' . md5(time());
 
         $this->object->setFileSet($fileSet);
@@ -96,6 +102,7 @@ class ActionTest extends \PHPUnit_Framework_TestCase
     }
     /**
      * @covers Sagres\Framework\FileSystem\Action::copyToFolder
+     * @covers Sagres\Framework\FileSystem\Action::copy
      * @expectedException Sagres\Framework\FileSystem\Exception\IOException
      * @depends testCopyToFolder
      */
@@ -113,6 +120,7 @@ class ActionTest extends \PHPUnit_Framework_TestCase
 
     /**
      * @covers Sagres\Framework\FileSystem\Action::copyToFolder
+     * @covers Sagres\Framework\FileSystem\Action::copy
      * @expectedException Sagres\Framework\FileSystem\Exception\InvalidPermissions
      * @depends testCopyToFolder
      */
@@ -131,6 +139,81 @@ class ActionTest extends \PHPUnit_Framework_TestCase
 
     }
 
+    /**
+     * @covers Sagres\Framework\FileSystem\Action::copyToFolder
+     * @covers Sagres\Framework\FileSystem\Action::copy
+     * @depends testCopyToFolder
+     */
+    public function testCopyToFolder_SourceFileNotFound()
+    {
+        $fileSet = new Set();
+        $fileSet->addSet(__DIR__ . '/../../../../fixtures/folder1');
+        $fileSet->addPath(__DIR__ . '/../../../../fixtures/not.found');
+        $folder = __DIR__ . '/../../../../fixtures/copy1';
+
+        $this->resetFolder($folder);
+
+        $this->object->setFileSet($fileSet);
+        $this->object->copyToFolder($folder);
+    }
+
+    /**
+     * @covers Sagres\Framework\FileSystem\Action::copy
+     * @expectedException Sagres\Framework\FileSystem\Exception\IOException
+     */
+    public function testCopy_originalFileIsFolder()
+    {
+        $fileSet = new Set();
+
+        $folder = __DIR__ . '/../../../../fixtures/copy1';
+
+        $this->object->copy($folder, $folder);
+    }
+
+    /**
+     * @covers Sagres\Framework\FileSystem\Action::copy
+     * @expectedException Sagres\Framework\FileSystem\Exception\NotFound
+     */
+    public function testCopy_originalFileIsNotFound()
+    {
+        $fileSet = new Set();
+
+        $folder = __DIR__ . '/../../../../fixtures/copy1/sadsadasdasD';
+
+        $this->object->copy($folder, $folder);
+    }
+
+    /**
+     * @covers Sagres\Framework\FileSystem\Action::copy
+     * @expectedException Sagres\Framework\FileSystem\Exception\InvalidPermissions
+     */
+    public function testCopy_originalFileIsNotReadable()
+    {
+        $fileSet = new Set();
+
+        $file = __DIR__ . '/../../../../fixtures/not.readable';
+        $file1 = __DIR__ . '/../../../../fixtures/copy/not.readable';
+
+        chmod($file, 0111);
+
+        $this->object->copy($file, $file1);
+    }
+
+    /**
+     * @covers Sagres\Framework\FileSystem\Action::copy
+     * @expectedException Sagres\Framework\FileSystem\Exception\NotFound
+     */
+    public function testCopy_DestinationFolderIsNotFound()
+    {
+        $fileSet = new Set();
+
+        $file = __DIR__ . '/../../../../fixtures/empty.file';
+        $file1 = __DIR__ . '/../../../../fixtures/xcsdafczxcxzcxzc/not.readable';
+
+
+        $this->object->copy($file, $file1);
+    }
+
 
     /**
      * @covers Sagres\Framework\FileSystem\Action::__construct
@@ -141,7 +224,48 @@ class ActionTest extends \PHPUnit_Framework_TestCase
         $o =new Action($fileSet);
 
         $this->assertSame($fileSet, $o->getFileSet());
-
     }
+
+    /**
+     * @covers Sagres\Framework\FileSystem\Action::mkdir
+     */
+    public function testMkdir()
+    {
+        $folder = __DIR__ . '/../../../../fixtures/foo';
+        $this->object->mkdir($folder);
+
+        $this->assertTrue(file_exists($folder));
+
+        rmdir($folder);
+    }
+    /**
+     * @covers Sagres\Framework\FileSystem\Action::mkdir
+     */
+    public function testMkdir_recursive()
+    {
+        $folder = __DIR__ . '/../../../../fixtures/foo/1';
+        $this->object->mkdir($folder);
+
+        $this->assertTrue(file_exists($folder));
+
+        rmdir($folder);
+        rmdir(dirname($folder));
+    }
+    /**
+     * @covers Sagres\Framework\FileSystem\Action::mkdir
+     */
+    public function testMkdir_folderAlreadyExists()
+    {
+        $folder = __DIR__ . '/../../../../fixtures/foo';
+        $this->object->mkdir($folder);
+
+        $this->assertTrue(file_exists($folder));
+
+        $this->object->mkdir($folder);
+
+        $this->assertTrue(file_exists($folder));
+        rmdir($folder);
+    }
+
 }
 ?>
